# api-contract/src/main/resources/static/brightlife_api_contract.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Bright-life Enhancement API
  description:
    "API documentation for Bright-life Enhancement project"
  license:
    name: "fronteersdev Inc"
    url: "https://fronteersdev.com/license"
servers:
  - url: http://localhost:7081
    description: Local test server
  - url: https://bright-life-enhancement.fronteers-dev.com/
    description: Dev test server
security:
  - BearerAuth: [ ]  # ⬅️ Apply JWT security globally to all endpoints
components:
  securitySchemes:
    BearerAuth:
      type: http
      description: This API uses JWT for authentication
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      explode: true
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      style: form
    limitParam:
      explode: true
      in: query
      name: limit
      required: false
      schema:
        format: int32
        maximum: 100
        type: integer
      style: form
...
    idParam:
      in: path
      name: idParam
      description: A unique id to update or fetch a record.
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Address:
      properties:
        id:
          type: integer
          format: int64
        streetName:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
      type: object
...
    GenderEnum:
      type: string
      enum:
        - Male
        - Female
        - Transgender
        - Declined
        - Non-Binary
        - Prefer Not To Say
    MaritalStatusEnum:
      type: string
      enum:
        - Single
        - Married
        - Divorced
        - Widowed
        - Separated
        - Domestic Partnership
        - Prefer Not To Say
...
    PaymentModeEnum:
      type: string
      enum:
        - Self Pay
        - Insurance Card
...
    PaginatedAppointments:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        itemsPerPage:
          type: integer
          example: 100
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
    Appointment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        patientId:
          $ref: "#/components/schemas/PatientId"
        isNew:
          type: boolean
        verificationStatus:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        gender:
          $ref: '#/components/schemas/GenderEnum'
        dob:
          $ref: "#/components/schemas/Date"
        phone:
          type: string
        email:
          $ref: "#/components/schemas/Email"
        address:
          $ref: '#/components/schemas/Address'
        appointmentType:
          type: string
          enum:
            - In-Person
            - Tele-Health
            - Any
        service:
          type: string
          enum:
            - Med Management
            - Individual Therapy
            - Group Therapy
            - Case Management
            - Psychiatric Rehab.
        appointmentDateTime:
          $ref: "#/components/schemas/DateTime"
        purpose:
          type: string
        paymentMethod:
          $ref: '#/components/schemas/PaymentModeEnum'
        insuranceName:
          $ref: '#/components/schemas/InsuranceNameEnum'
        insuranceNumber:
          type: string
    InsuranceNameEnum:
      type: string
      enum:
        - Medicare
        - Main Medicaid
        - BCBS Carefirst
        - UHC/Optum/Medicaid
        - Maryland Physicians Care
        - Amerigroup
...
    AppointmentSearch:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        dob:
          $ref: "#/components/schemas/Date"
        phone:
          type: string
        email:
          $ref: "#/components/schemas/Email"
        gender:
          $ref: "#/components/schemas/GenderEnum"
        maritalStatus:
          $ref: "#/components/schemas/MaritalStatusEnum"
        city:
          type: string
        state:
          type: string
        appointmentDateTime:
          $ref: "#/components/schemas/DateTime"
        paymentMode:
          $ref: '#/components/schemas/PaymentModeEnum'
...
    TimeSlot:
      type: object
      properties:
        date:
          $ref: "#/components/schemas/Date"
        daySlots:
          type: array
          items:
            type: string
    TimeSlots:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
...
    PatientId:
      type: string
      format: uuid
    Email:
      type: string
      #      pattern: "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+${5,200}"
      format: email
    DateTime:
      type: string
      format: date-time
    Date:
      type: string
      format: date
...
    Error:
      description: |

        **The below table describes custom error codes returned by various APIs currently**
        | Code      | error         | Message                     |
        | ----------| --------------|-----------------------------|
        | 400       | INPUT         | BAD REQUEST                 |
        | 404       | NOT_FOUND     | RESOURCE DOES NOT EXIST     |
        | 422       | UNPROCESSABLE | UNPROCESSABLE ENTITY        |
        | 401       | UNAUTHORIZED  | UNAUTHORIZED                |
        | 403       | FORBIDDEN     | FORBIDDEN                   |
        | 500       | PROCESSING    | INTERNAL SERVER ERROR       |
        | 500       | TIMEOUT       | TIMEOUT                     |
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
    Success:
      type: object
      required:
        - status
        - success
        - message
      properties:
        status:
          type: boolean
          default: false
        success:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      #      type: object
      description: |
        Bad Request
        The request body or parameters does not conform to the schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        Unauthorized.

        - Indicates the request does not contain a valid token to authenticate.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Forbidden.
            - Indicates the authenticated entity does not have the privilege to perform
              the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnProcessableEntity:
      description: |
        Unprocessable entity.
        - The request body may be well-formed, but either request payload or parameters failed semantic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: |
        Internal Server Error.
            - The server encountered an unexpected error while processing the request..
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /api/v1/appointments:
    post:
      summary: Submits An Appointment
      tags:
        - Appointment
      operationId: submitAppointment
      requestBody:
        description: Submit All Kinds of Form
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        "201":
          description: Appointment Submitted Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/appointments/search:
    post:
      summary:
        Searches and returns all Appointments matching all of the search criteria fields passed.
        The number of search results shall be constrained by the limit params in Query.
        Search criteria, applies an **AND** based search on all fields passed in the request.
        If the search criteria is empty/null, it returns list of appointments.
      tags:
        - Appointment
      operationId: listAppointments
      parameters:
        - description: The page number (starting from 1)
          $ref: "#/components/parameters/PageParam"
        - description: How many items to return at one time (max 100)
          $ref: "#/components/parameters/limitParam"
      requestBody:
        description:
          "Search criteria, applies an **AND** based search on all fields\
                \ passed in the request."
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentSearch"
      responses:
        "200":
          description: Appointment Records Fetched Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAppointments"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/appointments/slots:
    get:
      summary: Fetches Three Months Already Picked Appointment Slots From the Current Day
      tags:
        - Appointment
      operationId: getAppointmentSlots
      responses:
        "200":
          description: Already Picked Appointment Slots Retrieved Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSlots"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/appointments/{idParam}:
    get:
      summary: Fetches An Appointment By ID
      tags:
        - Appointment
      operationId: getAppointment
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: Appointment Retrieved Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/appointments/update:
    put:
      summary: Updates An Appointment
      tags:
        - Appointment
      operationId: updateAppointment
      requestBody:
        description: Submit All Kinds of Form
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        "201":
          description: Appointment Updated Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnProcessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...